Move classes `PDO` to directory `src/Driver/PDO`.
