+++ src/Driver/PDO/CommandPDOInterface.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Command;
+namespace Yiisoft\Db\Driver\PDO;
 
 use PDOStatement;
+use Yiisoft\Db\Command\CommandInterface;
 
 interface CommandPDOInterface extends CommandInterface
 {
