+++ src/Driver/PDO/ConnectionPDOPoolInterface.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Connection;
+namespace Yiisoft\Db\Driver\PDO;
 
 use PDO;
 use Throwable;
+use Yiisoft\Db\Connection\ConnectionInterface;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
