+++ src/Driver/PDO/PDOValueBuilder.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Pdo;
+namespace Yiisoft\Db\Driver\PDO;
 
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionInterface;
 
 /**
- * Class PdoValueBuilder builds object of the {@see PdoValue} expression class.
+ * Class PDOValueBuilder builds object of the {@see PDOValue} expression class.
  */
-class PdoValueBuilder implements ExpressionBuilderInterface
+final class PDOValueBuilder implements ExpressionBuilderInterface
 {
     public const PARAM_PREFIX = ':pv';
 
     public function build(ExpressionInterface $expression, array &$params = []): string
     {
-        $placeholder = (string) static::PARAM_PREFIX . count($params);
+        $placeholder = self::PARAM_PREFIX . count($params);
         $params[$placeholder] = $expression;
         return $placeholder;
     }
