+++ src/Schema/ColumnSchema.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Db\Schema;
 
 use PDO;
 use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Pdo\PdoValue;
+use Yiisoft\Db\Driver\PDO\PDOValue;
 use Yiisoft\Strings\NumericHelper;
 
 /**
@@ -103,7 +103,7 @@ class ColumnSchema
             && isset($value[1])
             && in_array($value[1], $this->getPdoParamTypes(), true)
         ) {
-            return new PdoValue((string) $value[0], $value[1]);
+            return new PDOValue((string) $value[0], $value[1]);
         }
 
         switch ($this->phpType) {
